add inputs for fill request body
