+++ src/pages/AcceptInvitation/pages/FillRequest/components/index.ts
@@ -1,0 +1,1 @@
+export { default as FillRequestForm } from './FillRequestForm'
