+++ src/pages/AcceptInvitation/pages/index.ts
@@ -1,2 +1,2 @@
 export { default as EmailVerification } from './EmailVerification'
-export { default as FillRequestForm } from './FillRequestForm'
+export { default as FillRequest } from './FillRequest'
